Change Validate to run all checks for type
The Validate method currently exits on the first instance of
Func<T, bool> that returns true. This makes it difficult for users to
test that a particular message is sent multiple times in response to
a single Handle.

This change enumerates over all matching invocations, before returning
true if any of the invocations was successful.