Change Validate to run all checks for type
Anybody using NServiceBus.Testing

ExpectSend only executes until the first check passes for the specified
type.

The Validate method currently exits on the first instance of check that
returns true. This makes it impossible for users to test that a
particular message is sent multiple times in response to a single
handle.